# SourceCode/front-end/logsmart/openapi.json
{
	"openapi": "3.1.0",
	"info": { "title": "back-end", "description": "", "license": { "name": "" }, "version": "1.1.0" },
	"paths": {
...
		"/auth/login": {
			"post": {
				"tags": ["Authentication"],
				"summary": "Authenticates a user and returns a JWT token.",
				"description": "# Errors\nReturns an error if credentials are invalid or if authentication fails.",
				"operationId": "login",
				"requestBody": {
					"content": {
						"application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } }
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Login successful",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } }
						}
					},
					"400": {
						"description": "Invalid request data",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"401": {
						"description": "Invalid credentials",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"429": {
						"description": "Rate limit exceeded",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				}
			}
		},
		"/auth/me": {
			"get": {
				"tags": ["Authentication"],
				"summary": "Retrieves the profile information of the currently authenticated user.",
				"description": "# Errors\nReturns an error if the user is not found or if there's a database error.",
				"operationId": "get_current_user",
				"responses": {
					"200": {
						"description": "Current user information",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/UserResponse" } }
						}
					},
					"401": {
						"description": "Unauthorized - invalid or missing token",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"404": {
						"description": "User not found",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				},
				"security": [{ "bearer_auth": [] }]
			}
		},
		"/auth/password/request-reset": {
			"post": {
				"tags": ["Authentication"],
				"summary": "Initiates a password reset process by sending an email.",
				"description": "# Errors\nReturns an error if the email is invalid or the request fails.",
				"operationId": "request_password_reset",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/RequestPasswordResetRequest" }
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Password reset email sent",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/PasswordResetResponse" }
							}
						}
					},
					"400": {
						"description": "Invalid email",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				}
			}
		},
		"/auth/password/reset": {
			"post": {
				"tags": ["Authentication"],
				"summary": "Resets a user's password using a valid reset token.",
				"description": "# Errors\nReturns an error if the token is invalid/expired or the operation fails.",
				"operationId": "reset_password",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/ResetPasswordRequest" }
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Password reset successfully",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/PasswordResetResponse" }
							}
						}
					},
					"400": {
						"description": "Password validation failed",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"401": {
						"description": "Invalid or expired token",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				}
			}
		},
		"/auth/profile": {
			"put": {
				"tags": ["Authentication"],
				"summary": "Updates the profile information of the currently authenticated user.",
				"description": "# Errors\nReturns an error if validation fails or the update operation fails.",
				"operationId": "update_profile",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/UpdateProfileRequest" }
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Profile updated successfully",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/UserResponse" } }
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				},
				"security": [{ "bearer_auth": [] }]
			}
		},
		"/auth/register": {
			"post": {
				"tags": ["Authentication"],
				"summary": "Registers a new company and its administrative user.",
				"description": "# Errors\nReturns an error if validation fails, the email already exists, or registration fails.",
				"operationId": "register_company_admin",
				"requestBody": {
					"content": {
						"application/json": { "schema": { "$ref": "#/components/schemas/RegisterRequest" } }
					},
					"required": true
				},
				"responses": {
					"201": {
						"description": "Company admin registered successfully",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } }
						}
					},
					"400": {
						"description": "Invalid request data",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"409": {
						"description": "Email already exists",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"429": {
						"description": "Rate limit exceeded",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				}
			}
		},
		"/auth/verify": {
			"post": {
				"tags": ["Authentication"],
				"summary": "Verifies the validity of a JWT token.",
				"description": "# Errors\nReturns an error if the token is invalid, expired, or if there's a database error.",
				"operationId": "verify_token",
				"requestBody": {
					"content": {
						"application/json": { "schema": { "$ref": "#/components/schemas/VerifyTokenRequest" } }
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Token is valid",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/JwtVerifyResponse" } }
						}
					},
					"401": {
						"description": "Invalid or expired token",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					},
					"404": {
						"description": "User not found",
						"content": {
							"application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"AuthResponse": {
				"type": "object",
				"required": ["token", "user"],
				"properties": {
					"token": { "type": "string" },
					"user": { "$ref": "#/components/schemas/UserResponse" }
				}
			},
...
			"ErrorResponse": {
				"type": "object",
				"required": ["error"],
				"properties": {
					"error": { "type": "string" },
					"fields": {
						"type": ["object", "null"],
						"additionalProperties": { "type": "array", "items": { "type": "string" } },
						"propertyNames": { "type": "string" }
					}
				}
			},
...
			"JwtVerifyResponse": {
				"type": "object",
				"required": ["email"],
				"properties": { "email": { "type": "string" } }
			},
...
			"LoginRequest": {
				"type": "object",
				"required": ["email", "password"],
				"properties": {
					"email": { "type": "string", "example": "admin@example.com" },
					"password": { "type": "string", "example": "SecurePass123!" }
				}
			},
...
			"PasswordResetResponse": {
				"type": "object",
				"required": ["message"],
				"properties": { "message": { "type": "string" } }
			},
...
			"RegisterRequest": {
				"type": "object",
				"required": [
					"email",
					"first_name",
					"last_name",
					"password",
					"company_name",
					"company_address"
				],
				"properties": {
					"company_address": { "type": "string", "example": "123 Main St, City, Country" },
					"company_name": { "type": "string", "example": "Example Corp" },
					"email": { "type": "string", "example": "admin@example.com" },
					"first_name": { "type": "string", "example": "John" },
					"last_name": { "type": "string", "example": "Doe" },
					"password": { "type": "string", "example": "SecurePass123!" }
				}
			},
...
			"RequestPasswordResetRequest": {
				"type": "object",
				"required": ["email"],
				"properties": { "email": { "type": "string", "example": "user@example.com" } }
			},
			"ResetPasswordRequest": {
				"type": "object",
				"required": ["token", "new_password"],
				"properties": {
					"new_password": { "type": "string", "example": "NewPassword123!" },
					"token": { "type": "string", "example": "reset-token-here" }
				}
			},
...
			"UpdateProfileRequest": {
				"type": "object",
				"required": ["first_name", "last_name"],
				"properties": {
					"first_name": { "type": "string", "example": "John" },
					"last_name": { "type": "string", "example": "Doe" }
				}
			},
...
			"UserResponse": {
				"type": "object",
				"required": ["id", "email", "first_name", "last_name", "role", "created_at"],
				"properties": {
					"branch_id": { "type": ["string", "null"] },
					"company_id": { "type": ["string", "null"] },
					"company_name": { "type": ["string", "null"] },
					"created_at": { "type": "string", "format": "date-time" },
					"email": { "type": "string" },
					"first_name": { "type": "string" },
					"id": { "type": "string" },
					"last_name": { "type": "string" },
					"oauth_picture": { "type": ["string", "null"] },
					"oauth_provider": { "type": ["string", "null"] },
					"profile_picture_url": { "type": ["string", "null"] },
					"role": { "$ref": "#/components/schemas/UserRole" }
				}
			},
			"UserRole": {
				"type": "string",
				"enum": ["logsmart_admin", "company_manager", "branch_manager", "staff"]
			},
...
			"VerifyTokenRequest": {
				"type": "object",
				"required": ["token"],
				"properties": { "token": { "type": "string", "example": "jwt-token-here" } }
			}
		},
		"securitySchemes": { "bearer_auth": { "type": "http", "scheme": "bearer" } }
	},
...
}